Depurados scripts para parametrizarlos

+++ utils/split_video.sh
@@ -2,6 +2,7 @@
 
 ## Analizar opciones de línea de comandos
 #OPTS=$(getopt -o c:h --long color:,help -n 'make_gif.sh' -- "$@")
+#OPTS=$(getopt -o r:h --long fps,help -n $0 -- "$@")
 OPTS=$(getopt -o h --long help -n $0 -- "$@")
 
 if [ $? -ne 0 ]; then
@@ -14,6 +15,7 @@ eval set -- "$OPTS"
 
 ## Inicializar variables
 #COLOR="0xB2D445"
+#FPS="25"
 HELP=false
 
 ## Procesar las opciones
@@ -23,8 +25,8 @@ while true; do
       HELP=true
       shift
       ;;
-#    -c | --color)
-#      COLOR="$2"
+#    -r | --fps)
+#      FPS="$2"
 #      shift 2
 #      ;;
     --)
@@ -61,14 +63,31 @@ do
 	echo
 	echo "FFMPEG $FILE a $PTH/$1 CROP: $5 desde $2 hasta $3 con $6 fps"
 #	sleep 4
-	ffmpeg -y -nostdin -avoid_negative_ts make_zero -ss "$2" -to "$3" -i "$FILE" -vf "crop=$5" -r "$6" "$PTH/$1.mp4" 2>&1
+	PARAM="-y -nostdin -avoid_negative_ts make_zero -ss $2 -to $3 -i $FILE"
 
-	[ -d /tmp/$1 ] && rm -r /tmp/$1
-	mkdir /tmp/$1
+	if [ "$6" == "false" ]; then
+		echo "NO HAY TRANSFORMACIÓN DE FPS......."
+	else
+		echo "HAY TRANSFORMACIÓN FPS......."
+		PARAM += "-r \"$6\"";
+
+	fi
+
+	echo "PARAM=$PARAM"
+
+	ffmpeg $PARAM -vf "crop=$5" "$PTH/$1.mp4" 2>&1 > /dev/null
+
+#ffmpeg -y -nostdin -avoid_negative_ts make_zero -ss "$2" -to "$3" -i "$FILE" -vf "crop=$5" -r "$6" "$PTH/$1.mp4" 2>&1
+
 	
-	echo ">>>>> FFMPEG para hacer transparencias de fotogramas en /tmp/$1"
-	ffmpeg -nostdin -i "$PTH/$1.mp4" -vf "colorkey=$4:0.1:0.0" -vsync 0 /tmp/$1/$1_%03d.png
-	echo ">>>>> CONVERT para reconstruir el gif animado en $PTH/$1-SR,gif"
-	convert -dispose background -delay 10 -loop 0 /tmp/$1/$1_*.png $PTH/$1-SR.gif
+	if [ "$4" != "false" ]; then
+		[ -d /tmp/$1 ] && rm -r /tmp/$1
+		mkdir /tmp/$1
+
+		echo ">>>>> FFMPEG para hacer transparencias de fotogramas en /tmp/$1"
+		ffmpeg -nostdin -i "$PTH/$1.mp4" -vf "colorkey=$4:0.1:0.0" -vsync 0 /tmp/$1/$1_%03d.png
+		echo ">>>>> CONVERT para reconstruir el gif animado en $PTH/$1-SR,gif"
+		convert -dispose background -delay 10 -loop 0 /tmp/$1/$1_*.png $PTH/$1-SR.gif
+	fi
 
 done < ${PTH}/${NAME}.cut
